backend/services/transcriber/requirements.txt
# FastAPI and web server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Audio processing
librosa>=0.10.0
soundfile>=0.12.0
pydub>=0.25.0

# Speech-to-Text
openai-whisper>=20231117
torch>=2.0.0
torchaudio>=2.0.0

# HTTP client and utilities
httpx>=0.25.0
requests>=2.31.0

# Data validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Logging
loguru>=0.7.0

# Environment
python-dotenv>=1.0.0
